oidc_provider/auth_perms/requirements.txt
ecdsa==0.16.0
email-validator==1.1.1
fastecdsa==2.1.5
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.9
Flask-Script==2.0.6
flask-shell-ipython==0.4.1
gevent==20.9.0
greenlet==0.4.17
Pillow==8.0.0
psycopg2-binary==2.8.6
py-healthcheck==1.10.1
python-engineio==3.13.2
python-usernames==0.3.1
qrcode==6.1
requests==2.24.0
pygit2==1.4.0
python-dotenv==1.0.0